Refactoring test items
- Rename token
- Add test create item
- Add test get item
- Add test send items
- Add test get list of items
- Fix answer code

+++ api_app/views.py
@@ -6,10 +6,9 @@ import jwt
 from flask import abort, jsonify, make_response, request, url_for
 from marshmallow import ValidationError
 
+from . import app, db
 from .models import Item, User
 from .schemes import ItemSchema, NewUserSchema, UserSchema
-
-from . import app, db
 
 
 def token_required(f):
@@ -20,13 +19,17 @@ def token_required(f):
             token = request.headers["x-access-tokens"]
 
         if not token:
-            return jsonify({"message": "a valid token is missing"})
+            return make_response(jsonify({"message": "Token is missing"}), 403)
 
         try:
             data = jwt.decode(token, app.config["SECRET_KEY"], algorithms=["HS256"])
             current_user = User.query.filter_by(username=data.get("username")).first()
         except jwt.exceptions.InvalidTokenError:
-            return jsonify({"message": "Token is invalid"})
+            return make_response(jsonify({"message": "Token is invalid"}), 403)
+        if not current_user:
+            return make_response(
+                jsonify({"message": "Тoken does not belong to any user"}), 403
+            )
         return f(current_user, *args, **kwargs)
 
     return decorator
@@ -37,12 +40,17 @@ def internal_server_error(e):
     return make_response(jsonify({"message": "Internal Server Error"}), 500)
 
 
+@app.errorhandler(404)
+def internal_server_error(e):
+    return make_response(jsonify({"message": "URL not Found"}), 404)
+
+
 @app.errorhandler(400)
 def internal_server_error(e):
     return make_response(
         jsonify(
             {
-                "message": "400 Bad Request: The browser (or proxy) sent a request that this server could not understand."
+                "message": "400 Bad Request: The browser (or proxy) sent a request that this server could not understand"
             }
         ),
         400,
@@ -63,8 +71,6 @@ def index(user):
 def create_item(user):
     item_schema = ItemSchema()
     json_data = request.get_json()
-    if not json_data:
-        return make_response(jsonify({"message": "No input data provided"}), 200)
     try:
         data = item_schema.load(json_data)
     except ValidationError as err:
@@ -80,12 +86,16 @@ def create_item(user):
 @token_required
 def delete_item(user, id):
     item = Item.query.get(id)
-    try:
-        db.session.delete(item)
-        db.session.commit()
-        return make_response(jsonify({"item": f"Item: { item.name } deleted"}), 200)
-    except error as err:
-        return make_response(jsonify({"message": f"{ err }"}), 422)
+    if not item:
+        return make_response(jsonify({"message": "No item with such id"}), 422)
+    check_user = User.query.filter(User.items.any(Item.id == id)).first()
+    if not check_user == user:
+        return make_response(
+            jsonify({"message": "This user can,t delete this item"}), 403
+        )
+    db.session.delete(item)
+    db.session.commit()
+    return make_response(jsonify({"item": f"Item: { item.name } deleted"}), 200)
 
 
 @app.route("/api/v1/send", methods=["POST"])
@@ -93,21 +103,27 @@ def delete_item(user, id):
 def send_item(user):
     json_data = request.get_json()
     new_user_schema = NewUserSchema()
-    if not json_data:
-        return make_response(jsonify({"message": "No input data provided"}), 200)
     try:
         data = new_user_schema.load(json_data)
     except ValidationError as err:
         return make_response(jsonify({"message": f"{ err.messages }"}), 422)
     item_id, new_username = data["item_id"], data["new_username"]
-    is_user_has_item = User.query.filter(
-        User.username == new_username, User.items.any(Item.id == item_id)
-    ).all()
-    if is_user_has_item:
-        return make_response(jsonify({"message": "User already has this item"}), 200)
+    check_users = User.query.filter_by(username=new_username).all()
+    if not check_users:
+        return make_response(jsonify({"message": "No destination user"}), 422)
     check_items = Item.query.filter_by(id=item_id).all()
     if not check_items:
-        return make_response(jsonify({"message": "No item with such id"}), 200)
+        return make_response(jsonify({"message": "No item with such id"}), 422)
+    user_items = User.query.filter(
+        User.username == user.username, User.items.any(Item.id == item_id)
+    ).all()
+    if not user_items:
+        return make_response(jsonify({"message": "Item not belong to user"}), 403)
+    new_user_items = User.query.filter(
+        User.username == new_username, User.items.any(Item.id == item_id)
+    ).all()
+    if new_user_items:
+        return make_response(jsonify({"message": "User already has this item"}), 422)
     move_token = jwt.encode(
         {
             "item_id": item_id,
@@ -128,14 +144,23 @@ def get_item(user, move_token):
             move_token, app.config["SECRET_KEY"], algorithms=["HS256"]
         )
     except jwt.exceptions.InvalidTokenError:
-        return make_response(jsonify({"message": "Token is invalid"}), 200)
+        return make_response(jsonify({"message": "Token in url is invalid"}), 422)
     item_id, new_username = move_token_data.get("item_id"), move_token_data.get(
         "new_username"
     )
     if not user.username == new_username:
-        return make_response(jsonify({"message": "Another user token"}), 200)
-    item = Item.query.get(item_id)
+        return make_response(jsonify({"message": "Another user token"}), 403)
     new_user = User.query.filter_by(username=new_username).one()
+    item = Item.query.get(item_id)
+    if not item:
+        return make_response(jsonify({"message": "Item is not found"}), 422)
+    user_items = User.query.filter(
+        User.username == new_username, User.items.any(Item.id == item_id)
+    ).all()
+    if user_items:
+        return make_response(
+            jsonify({"message": "User already has this item or reuse url"}), 422
+        )
     item.user_id = new_user.id
     db.session.commit()
     result = item_schema.dump(Item.query.get(item.id))
@@ -145,12 +170,7 @@ def get_item(user, move_token):
 @app.route("/api/v1/user/registration", methods=["POST"])
 def create_user():
     user_schema = UserSchema()
-    try:
-        json_data = request.get_json()
-    except Exception as err:
-        return make_response(jsonify({"message": f"{ err }"}), 400)
-    if not json_data:
-        return make_response(jsonify({"message": "No input data provided"}), 200)
+    json_data = request.get_json()
     try:
         data = user_schema.load(json_data)
     except ValidationError as err:
@@ -169,8 +189,6 @@ def create_user():
 def login_user():
     user_schema = UserSchema()
     json_data = request.get_json()
-    if not json_data:
-        return make_response(jsonify({"message": "No input data provided"}), 200)
     try:
         data = user_schema.load(json_data)
     except ValidationError as err:
@@ -180,13 +198,13 @@ def login_user():
     if not user:
         return make_response(jsonify({"message": "User is not exist"}), 422)
     if user.verify_password(password):
-        web_token = jwt.encode(
+        auth_token = jwt.encode(
             {
                 "exp": datetime.utcnow()
-                + timedelta(seconds=app.config["WEB_TOKEN_PERIOD_EXPIRE_SECONDS"]),
+                + timedelta(seconds=app.config["AUTH_TOKEN_PERIOD_EXPIRE_SECONDS"]),
                 "username": user.username,
             },
             app.config["SECRET_KEY"],
         )
-        return make_response(jsonify({"user": {"web_token": web_token}}), 200)
+        return make_response(jsonify({"user": {"auth_token": auth_token}}), 200)
     return make_response(jsonify({"message": "Wrong password!"}), 403)
